finished disassembler

+++ include/disassembler.h
@@ -8,4 +8,4 @@
 using namespace std;
 
 // disassemble returns array of strings of memory disassembly
-vector<string> disassemble(uint8_t* mem, uint16_t size, const InstructionSet& instset);
+vector<string> disassemble(uint8_t* mem, uint32_t size, const InstructionSet& instset);
